requirements.txt
anthropic==0.43.0
google-api-python-client==2.114.0
google-auth==2.27.0
jinja2==3.1.3
requests==2.31.0
sendgrid==6.11.0
python-dotenv==1.0.1
